New Forecasting Logic Added

+++ CloudERP/Models/Forecasting/ForecastPrediction.cs
@@ -5,6 +5,6 @@ namespace CloudERP.Models.Forecasting
     public class ForecastPrediction
     {
         [ColumnName("Score")]
-        public float PredictedValue { get; set; }
+        public double PredictedValue { get; set; }
     }
 }
